requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-api-core==2.25.2
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-generativeai
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
gunicorn==25.1.0
uvicorn[standard]==0.34.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
openai==2.26.0
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
whitenoise==6.12.0
celery==5.4.0
redis==5.2.1
django-redis==5.4.0
django-celery-results==2.5.1
django-storages
boto3
